PascalVOC: Use sample list files instead annotations to build sample lists
Pascal VOC contains sample list files that specify which files belong to
particular sets. Previously the code created the training, validation, and test
sets by just listing the annotation XML files in appropriate directories. It
turns out that these do not have a one to one correspondence in the VOC 2012
trainval dataset. This change limits the traininig set to whatever is specified
in the sample list file and uses the remaining samples and annotations as
a validation set.

Suggested by Changjo in PR #12.

Fixes #12.